Realized that things had to change for web
So, I think that what we came up with is good because it helped me
work through the logic of what we need. I think we still need all the
same functions. But, I need to change the logic to interact with SQLite.
Or to use an ORM like SQLAlchemy.